Update README.md
The Addis Ababa ECD Monthly report excel has 5 tabels with 49 columns to be filled with data this tables are separated by headers, so you just cant copy and past to an excel and get a sum. The data collector will have to add this values from 49 columns found on each excel to one reportable. They use to write each values from 12 excels and add them on paper get the final total. This python code will extract the data from all the different excels and put out a single reportable. this is done with panads data frame and excel packages for python. all the will do is rename the file as "report", like report(1),report(2) upload the files to Google Collab for free and run this code.